added file to the models directory, fixed the test
